Move test_point_ctad
